+++ compiler-tests/src/test/data/diagnostic/bindingcontainer/IllegalTypes.kt
@@ -22,4 +22,15 @@ sealed interface SealedInterface
 class Containing {
   <!BINDING_CONTAINER_ERROR!>@BindingContainer<!>
   inner class InnerClass
+
+  fun example() {
+    val anonymous =
+    <!BINDING_CONTAINER_ERROR!>@BindingContainer<!>
+      object : Any() {
+
+      }
+
+    <!BINDING_CONTAINER_ERROR!>@BindingContainer<!>
+    class LocalClass
+  }
 }