+++ CHANGELOG.md
@@ -23,6 +23,8 @@ Changelog
 - **Fix**: Don't report duplicate bindings if injectors for both a parent and child class are present on a graph.
 - **Fix**: Look up correct target class ID for computed member injectors in `BindingLookup`.
 - **Fix**: Don't allow binding containers to be `inner` classes.
+- **Fix**: Don't allow binding containers to be local classes.
+- **Fix**: Don't allow binding containers to be anonymous objects.
 
 0.6.10
 -----